django-rest-framework-homework/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
django==4.2.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
